\name{grip}
\alias{grip}
\docType{data}
\title{The hand grip strength data}
\description{
The data is a subset (only boys) from the data analysyed by  Cohen \emph{et al.} (2010). 
}
\usage{data("grip")}
\format{
  A data frame with 3766 observations on the following 2 variables.
  \describe{
    \item{\code{age}}{the age of the participant}
    \item{\code{grip}}{the  handgrip strength}
  }
}
\details{
 Cohen \emph{et al.} (2010)  analysed the of hand grip (HG) strength in relation to gender and age in
English schoolchildren. Here there are  3766  observations of the boys. 
}
\references{
 Cohen, D.D.,Voss, C.,  Taylor, M.J.D., Stasinopoulos, D.M., Delextrat, A. and Sandercock, G.R.H. (2010) Handgrip strength in English schoolchildren, \emph{Acta Paediatrica}, \bold{99}, 1065-1072.
}
\examples{
data(grip)
}
\keyword{datasets}
