\name{rent99.polys}
\alias{rent99.polys}
\docType{data}
\title{
The boundaries file for  Munich rent data from the 1999 survey.
}
\description{
The boundaries files of of 1999 Munich survey.
}
\usage{data(rent99.polys)}
\format{
This data frame contains the boundaries of the Munich data. 
}
\details{
See Fahrmeir et. al., (2013) page 5, for more details about the data.
}
\source{
Thanks to Thomas Kneib who provide us with the data.
}
\references{
Fahrmeir, Ludwig and Kneib, Thomas and Lang, Stefan and Marx, Brian (2013) \emph{Regression: models, methods and applications}, Springer.
}
\examples{
data(rent99.polys)
## library(gamlss.spatial); draw.polys(rent99.polys)
}
\keyword{datasets}
