\name{triceps}
\alias{triceps}
\docType{data}
\title{Triceps Skinfold in Gambian females}
\description{Data from an anthropometry survey of 892 girls and women aged 0-50 years from three Gambian villages in West Africa. 
}

\usage{data("triceps")}
\format{
  A data frame with 892 observations on the following 2 variables.
  \describe{
    \item{\code{Age}}{the age in years}
    \item{\code{tricep}}{triceps skinfold in mm}
  }
}
\details{
Data from an anthropometry survey of 892 girls and women aged 0-50 years from three Gambian villages in West Africa first analysed by Cole and Green (1992).
}
\source{
 Cole and Green (1992)
}
\references{
Cole,  T. J. and Green,  P. J. (1992) Smoothing reference centile curves: the LMS method and penalized likelihood, \emph{Statist. Med.} \bold{11}, 1305--1319
}
\examples{
data(triceps)
head(triceps)
}
\keyword{datasets}
